fix(gfx/spritesys): use ptrdiff_t for Shadow_565_32 pixel stride arithmetic
UBSan during autoplay flagged:
  runtime error: addition of unsigned offset to 0x... overflowed to 0x...
at SpriteLow565.cpp:1822 inside Sprite::__Shadow_565_32.

The line was:
  dest.b_ptr += dest_inc * sizeof(Pixel32);

dest_inc is sint32 (can be negative for reversed/flipped draws), but
sizeof yields size_t. The multiplication promotes dest_inc to unsigned,
so -1 becomes 0xFFFFFFFF and the product becomes a huge unsigned offset
that wraps the pointer arithmetic when added to dest.b_ptr.

Cast both operands through ptrdiff_t so the multiplication stays signed,
producing a negative byte delta that pointer addition handles correctly.

+++ ctp2_code/gfx/spritesys/SpriteLow565.cpp
@@ -1819,6 +1819,9 @@ inline void Sprite::__Shadow_565_32(PixelAddress &dest,
 	memcpy(&shadowedPixel, dest.b_ptr, sizeof(Pixel32));
 	shadowedPixel = pixelutils_Shadow32_565(shadowedPixel);
 	memcpy(dest.b_ptr, &shadowedPixel, sizeof(Pixel32));
-  	 dest.b_ptr += dest_inc * sizeof(Pixel32);
+  	 // Use ptrdiff_t so negative dest_inc (reversed/flipped draws) stays
+  	 // signed through the multiply.  size_t promotion would underflow the
+  	 // pointer arithmetic (UBSan-flagged during autoplay).
+  	 dest.b_ptr += (ptrdiff_t)dest_inc * (ptrdiff_t)sizeof(Pixel32);
   }
 }
